Remove:eslint, Rename:디렉토리구조변경
