--- client/.eslintrc.json
@@ -1,34 +1,0 @@
-{
-    "root": true,
-    "extends": [
-        "eslint:recommended",
-        "plugin:react/recommended",
-        "plugin:@typescript-eslint/recommended",
-        "plugin:@typescript-eslint/eslint-recommended",
-        "plugin:prettier/recommended",
-        "prettier"
-    ],
-    "plugins": ["@typescript-eslint", "prettier"],
-    "parser": "@typescript-eslint/parser",
-    "parserOptions": {
-        "project": "./tsconfig.json",
-        "ecmaFeatures": {
-            "jsx": true
-        },
-        "ecmaVersion": 2022,
-        "tsconfigRootDir": __dirname,
-        "sourceType": "module"
-    },
-
-    "rules": {
-        "@typescript-eslint/strict-boolean-expressions": [
-            2,
-            {
-                "allowString": false,
-                "allowNumber": false
-            }
-        ],
-        "react/react-in-jsx-scope": "off"
-    },
-    "ignorePatterns": ["src/**/*.test.ts", "src/frontend/generated/*"]
-}
