+++ client/package.json
@@ -14,7 +14,6 @@
         "@types/react": "^18.0.0",
         "@types/react-dom": "^18.0.0",
         "axios": "^0.27.2",
-        "eslint-plugin-react": "^7.31.8",
         "react": "^18.2.0",
         "react-dom": "^18.2.0",
         "react-router-dom": "6",
@@ -42,13 +41,6 @@
         ]
     },
     "devDependencies": {
-        "@types/styled-components": "^5.1.26",
-        "@typescript-eslint/eslint-plugin": "^5.38.0",
-        "@typescript-eslint/parser": "^5.38.0",
-        "eslint": "^8.23.1",
-        "eslint-config-prettier": "^8.5.0",
-        "eslint-plugin-import": "^2.26.0",
-        "eslint-plugin-prettier": "^4.2.1",
         "prettier": "2.7.1"
     }
 }
